REFACTOR move PropertiesSidebar component
